Crate su3

su3

En-/decoder for the SU3 file format used by I2P for sending reseed information, updates and more

Format specification

License

This crate is licensed under the MIT license

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, shall be licensed as MIT, without any additional terms or conditions.

Examples

let (_, parsed_su3) = Su3::from_bytes((su3_data, 0)).expect("Failed to parse SU3 file");
let content = parsed_su3.content().expect("Failed to decompress content");

Re-exports

pub use deku;

Structs

Su3

Typed representation of an SU3 file

Enums

ContentType

Content type

FileType

File type

SignatureType

Signature type

Constants

MIN_VERSION_LENGTH

Minimum length of the version field